+++ src/types.ts
@@ -1,6 +1,6 @@
 export type Args = string[];
 export type TemplateTypeAnswer = 'reuseable' | 'withTest' | 'custom';
-export type CustomTemplateChoice = 'test' | 'story';
+export type CustomTemplateChoice = 'test' | 'story' | 'types' | 'style';
 export interface PromptAnswers {
   componentName: string;
   template: TemplateTypeAnswer;