+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for custom template
+- Support for custom template including test, types, style and component
+- Added types to Component with test premade template
 
 ### Changed
 
-- Added types for the Component + test premade template
+- Updated strings to use single quote
 
 ## 0.1.1 - 2022-10-22